dom/base/test/chrome/test_bug1063837.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=206691
-->
<window title="Mozilla Bug 1063837"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1063837"
     target="_blank">Mozilla Bug 1063837</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

  /** Test for Bug 1063837 **/
  SimpleTest.waitForExplicitFinish();

  addLoadEvent(function() {
    var xhr = new XMLHttpRequest();
    xhr.open("GET", location, false);
    xhr.onload = function() {
      ok(xhr.responseXML, "We should have response content!");
      var principal = xhr.responseXML.nodePrincipal;
      ok(principal.URI.schemeIs("moz-nullprincipal"), "The response document should have a null principal");
      SimpleTest.finish();
    }
    xhr.send();
  });
  ]]></script>
</window>
